Disable failed access auditd logging
By default, the security role enables audit logging for failed file
accesses. This causes lots of log lines due to the way that rsyslog
looks for existing log files on the system.

This patch disables the auditd rule by default and adds documentation
for deployers so they know how to opt-in to the change.

Release notes are included.

Closes-bug: 1577448

Change-Id: I9ce4a208f5b9f28a1f317cb25a8114b902f5cabb