Add check/audit to gate testing
Checking or auditing an existing environment is a first class feature
in the security role, but we're not currently checking that feature
in the gate. A few users of the role have come forward with some
bugs around the check mode functionality and we should test this
more regularly.

This patch adds a quick audit check right before the functional test runs.
This adds about 30-60 seconds onto the gate test, but it should help
with catching these bugs.

Change-Id: I5f4adc292f027b2eb7429be843c167d152c0169d