Restart auditd after running augenrules
The augenrules command joins together all of the audit rules from
rules.d and it is run any time the audit rules template changes. However,
the augenrules handler didn't actually restart auditd to apply the
changes to the system.

This patch fires off the auditd restart handler anytime the augenrules
handler is notified.

Closes-bug: 1590916

Change-Id: Ice83fe17ebb0e9edff9da897e435ae96c1778580