Initial docs scaffolding for RHEL 7 STIG
This patch updates the documentation and the sphinx extension to
write the docs for the RHEL 7 STIG content. The RHEL 7 content
is still hidden for now since it will be under active development
for a while.

Special note:
There are some typos in version 0.2 of the XML for the STIG and those
typos exist in the automatically generated documentation, too.
A new version of the STIG XML is due out very soon which should
correct these.

Implements: bp/security-rhel7-stig
Change-Id: I70dcebdd45ca65e5ae7734c19ee8523fba581e18