[Docs] Declutter controls listing
The listing of available STIG controls is fairly cluttered due to the
fix/check sections. That content comes from the STIG XML and the content
is under flux as the RHEL 7 STIG is getting released.

This patch removes the fix/check sections in the hopes of linking to the
STIG Viewer website soon. It also adds a severity marker to each STIG
listing so that readers better understand each control.

Change-Id: Ie44c2eaeab2b1faa069468ec5b2c4d63389d47ad