Update and clean up run_tests.sh
The run_tests.sh script fails when it is run multiple times on CentOS.
The `bindep` run returns an empty list of packages and then `yum`
exits with an error since no packages were provided to install.

This patch checks the length of the `bindep` output and skips the `yum`
installation when the package list is empty.

The patch also cleans up some of the old cruft left over from previous
scripts and avoids repetition.

Change-Id: Ibe4d0fd9d608dc725c354723143e60c89cd99b4b