Make umask change opt-in
Changing the default umask causes issues with OpenStack-Ansible
deployments in roles where directories are created without a mode
specified. It also may surprise some users on non-OpenStack systems
who expect the default umask to match the default from the OS.

This patch makes the change an opt-in change and it updates the
documentation to reflect that.

Related-bug: 1656003
Change-Id: I0931a34b1114e3a57e0eb5914124eed589ded541