Restore RHEL 6 STIG content gating
When the role was switched to use RHEL 7 STIG content as the default,
the RHEL 6 STIG content was no longer being gated. This patch adds a
new test environment, `func_rhel6` and replaces the `func_rhel7`
environment with testing of the RHEL 6 STIG content. This is only
temporary and the `func_rhel7` environment should be removed as soon
as upstream gating is updated to use the new `func_rhel6` environment.

Change-Id: I9f91dac8ecadd3b791954f18e0607403ff147876