Move test plugins into security role
The greaterthan/lessthan test plugins were added into
openstack-ansible-plugins a month ago, but they are needed in the
security role to ensure that the role can run on a system without
openstack-ansible-plugins installed.

The test plugins in this role should be removed once jinja 2.9.5 is
available for use in OpenStack (likely after Ocata release).

Closes-bug: 1660060
Change-Id: Ic86dc9a9e8ae8d03fe52a7c3fc52e580b9936711