tasks: rhel7stig: aide: Fix conditionals for Ubuntu exclusions
The ansible_os_family returns 'Debian' instead of 'Ubuntu' on Ubuntu
distributions so the conditional was never true and as such the task was
never executed. We fix the conditional to include the correct family and
also check on whether the configuration directory exists which is
similar to what the other tasks are doing. Finally, we fix the 'dest'
option in the 'blockinfile' module since {{ aide_conf }} is a directory
but we should use the configuration file instead.

Change-Id: I6b8101db57d27dd3ae38dd36cd9875d165e1afa5